grrenthumb-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
scikit-learn==1.3.2
opencv-python==4.8.1.78

# NEW: Add these for soil analysis
opencv-python==4.8.1.78
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3

# Optional: For better image processing
scipy==1.11.4